Split `stringFrom` and `stringTo` to separate header files
